Bug 7168 - MAC Learning from ARP to be allowed on Ext-Interfaces
An earlier fix to make sure that Neutron configured IPs are not
learned here https://git.opendaylight.org/gerrit/#/c/48054/, has
broken the ability to learn MAC Addresses on external network
via external interfaces.

With this fix, we open the gate to allow MAC learning on all
packets coming from external interface.  This is not the best
possible fix for the problem, but this tentatively allows
the openstack use-case of external network access to continue to
be functional.

Closes-Bug: 7168
Change-Id: Ief6f8bfe26111e07e280a9245a44032a8f32f09d
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>