Update patch set 1
Patch Set 1:

Suraj,

Just a general question before I start reviewing - a bit hard to see because of the refactoring into a new file - what is the bug fix here? Is there an open bug for this/do we want this fix in Boron (I assume so)?

I would assume some of the data was previously written into operational if it doesn't survive reboot, but would be great to get a summary of what changes in the logic, preferably in the commit message.. thanks!

Patch-set: 1
Label: Verified=0