Bug 7119: gw arp didn't resolve
Unbind service was not done while removing an external network
Fix - add vpn interface to the operational models for the external nw.

VpnInterface of the provider network port for flat/vlan network has no associated adjacencies but
still needs to be created in order to bind L3VPN service to traffic originated from ext-net.
As a result of not having VpnInterface in OPER DS we discovered there are leftovers in the service
bindings DS after VpnInterface removal (on ext-net delete).
An attempt to create new ext-net will fail to rebind to a new L3VPN service as this interface is already bound.
This fix will allow to clean up the service binding leftovers after ext-net delete operation. 

Change-Id: I933052538e7ccd4b90ce5e499e56affdd83219c7
Signed-off-by: Olga Schukin <olga.schukin@hpe.com>