Update patch set 1
Patch Set 1: -Code-Review

> (1 comment)

Vivek, this is the VpnInterface for the provider network port for flat/vlan network. It has no associated adjacencies but still needs to be created in order to bind L3VPN service to traffic originated from ext-net.
As a result of not having VpnInterface in OPER DS we discovered there are leftovers in the service bindings DS after VpnInterface removal (on ext-net delete).
An attempt to create new ext-net will fail to rebind to a new L3VPN service as this interface is already bound.
This fix will allow to clean up the service binding leftovers after ext-net delete operation. 

BTW - we are thinking about adding subnetRoute on this interface to resolve PNF addresses but I think this will be only for Carbon.

Patch-set: 1
Reviewer: Gerrit User 5802 <5802@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0