Update patch set 11
Patch Set 11:

> Hi Alon,
 > 
 > The review https://git.opendaylight.org/gerrit/#/c/48054/ already
 > addressed most part of the problem that this review attempts to
 > solve..
 > 
 > This has been lying around for a while that I need to revisit this
 > again to see what benefit this would offer over #48054.
 > 
 > Vivek

Thanks Vivek.

Vivek, I was worried that things being stored only in operational are cleared after reboot.
I assumed this review moves several datastores into config so they are also persistent after reboot - is this not the case?

If there's not much benefit of it I agree but let's revisit the benefits :)

Patch-set: 11