Update patch set 24
Patch Set 24:

> Peri had commented on this earlier in mail so adding those back
 > here.
 > The current ARP responder code programs table 81 flow entries for
 > each interfaces so that ARP response can be unicasted to interface
 > on which ARP request is received. It just uses ifaceMgrRpcService.getEgressActionsForInterface(...)
 > which returns [load:0->NXM_OF_IN_PORT[],load:<lporttag>->NXM_NX_REG6[],resubmit(,220)
 > ] for actions in table 81 flow entries. This is same as how ELAN
 > programs known DMAC table flow entries.  So I don’t see a need for
 > table 81 flow entries to do packet-out on ELAN DMAC table (51).

Achuth I saw that response - this just strengthens the fact that there is a "duplication" here, and since it results in a large number of flows I was wondering if it can be avoided. If there are actual benefits of this approach we can keep it this way, but otherwise it might be a good refactoring - was wondering what are Vivek/Vishal's thoughts on this.

Patch-set: 24