Spec for VNI based L2 switching and L3 router forwarding
1. Added specification document for VNI based L2 switching and L3 router
forwarding
2. Linked spec to specs/index

Today, for VXLAN enabled networks by the tenant, the labels are generated by L3
forwarding service and used.

This feature realizes VNI (Virtual Network Identifier) based L2 switching and
L3 router forwarding.
In doing so, it eliminates the usage of LPortTags, ELAN tags and MPLS labels
for realizing the same on the wire. Instead, it makes use of VNIs supplied by
the OpenStack using tenant.

In the nodes themselves, the LportTag will continue to be used when
pushing the packet from ELAN/L3VPN pipeline towards the VM as ACLService
continues to use LPortTags.

Simiarly ELANTags will continue to be used for handling broadcast packets
both within the Node for broadcasts generated within that node and also
for broadcasts that land from another node to the current node via
INTERNAL VXLAN Tunnels..

As stressed, the wire will carry VXLAN VNIs as dictated by the tenant's
Openstack configuration.

Change-Id: I708243b75fb83ce94064b7214af2709ee82914f4
Signed-off-by: Abhinav Gupta <abhinav.gupta@ericsson.com>