WiP: Spec for VNI based L2 switching, L3 forwarding and NATing
1. Added specification document for VNI based L2 switching, L3
forwarding and NATing
2. Linked spec to specs/index

This spec is a first step in the direction of enforcing datapath
semantics that uses tenant supplied VNI values on VXLAN Type networks
created by tenants in Openstack Neutron.

Earlier the datapath semantics for VXLAN traffic from tenants were
realized by using either LportTags (or) MPLS Labels being carried as
VNI values, even though the networks created by the tenant were of
type VXLAN.

Please note that implementation of Spec DOES NOT COMPLETELY remove
usage of LPortTags (or) ElanTags (or) MPLS Labels.

It retains the use of MPLSLabels for realizing an L3 BGPVPN
(regardless of type of networks put into such BGPVPN that may include
networks of type VXLAN) both on the wire and within the OVS Datapaths.

It also retains the use of LPortTags within the OVS Datapath of a
hypervisor, for streaming traffic to the right virtual endpoint on that
hypervisor (note: not on the wire).

It retains the use of ELANTags within the OVS Datapath to handle
local/remote L2 broadcasts (note: not on the wire)

On the cloud data network wire, all the VXLAN traffic for basic L2
Switching within a VXLAN network and L3 Forwarding across VXLAN-type
networks using Routers will use tenant supplied VNI values for such
VXLAN networks.

Complete removal of use of LPortTags and MPLS Labels for VXLAN-type
networks has large scale design/pipeline implications and thus need to
be attempted as future initiatives via respective specs.

Change-Id: I708243b75fb83ce94064b7214af2709ee82914f4
Signed-off-by: Abhinav Gupta <abhinav.gupta@ericsson.com>
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>