Spec for L3 forwarding for routers using LPort Tags
1. Added specification document for L3 forwarding for routers using LPort Tags
2. Linked spec to specs/index

For VXLAN enabled networks by the tenant, the labels are generated by L3
forwarding service and used. This implementation will completely remove
the reliance on using labels for L3 forwarding within the DC. Instead,
it will re-use the LPortTags (which are unique for every port in the
cloud) to provide L3 forwarding semantics in the datapath.

Change-Id: I708243b75fb83ce94064b7214af2709ee82914f4
Signed-off-by: Abhinav Gupta <abhinav.gupta@ericsson.com>