Update patch set 25
Patch Set 25:

> Hi Sam/Alon,
 > 
 > I could see only the same 6 testcases failed, which require
 > Venkat's fix to land and then to be re-run.
 > 
 > This time around the FloatingIP test cases passed too, indicative
 > of a timing issue.
 > 
 > Vivek
 > 
 > 
 > Vivek

Vivek, in general I'm ok with this as it is not breaking any current functionality - two question before +1 though:

a. Could you please respond to my comment on this gerrit about duplication of per-lport rules in ARP responder, while this is already done in ELAN tables? (Copying below)
b. Venkat's fix is still pending unfortunately - I'm guessing the tests failing with permission denied will pass. Note there were two tests that still timed out in the latest run, could you have a look?

"Achuth I saw that response - this just strengthens the fact that there is a "duplication" here, and since it results in a large number of flows I was wondering if it can be avoided. If there are actual benefits of this approach we can keep it this way, but otherwise it might be a good refactoring - was wondering what are Vivek/Vishal's thoughts on this"

Patch-set: 25