Update patch set 25
Patch Set 25:

Vivek, in general I'm ok with this as it is not breaking any current functionality - two question before +1 though:

a. Could you please respond to my comment on this gerrit about duplication of per-lport rules in ARP responder, while this is already done in ELAN tables? (Copying below)

[VIVEK]   This is a fallout of BGPVPN re-using the L3Forwarding service.
BGPVPN in openstack allows a bunch networks that donot have gateway ports created to be put in.  Now since BGPVPN needs to provide L3VPN connectivity, it has to shunt some MACAddress for ARP Request for subnet gateway ips.  

Since there is no gateway port available from Openstack for the networks in the BGPVPN, we have been historically using the MACAddress of the connected port (the port that connects the VM to the OVSDatapath) for responding to the ARP requests for Gateway IPs, which will then enable IP packets to flow through thereby providing L3 Connectivity.

In order to retain that behaviour, we have to program per lport rules into the ARP responder table.



b. Venkat's fix is still pending unfortunately - I'm guessing the tests failing with permission denied will pass. Note there were two tests that still timed out in the latest run, could you have a look?

I already looked at the two timeout reported first thing today.  When I browsed through the flow rules, all of them were intact for 100.0.0.3 and 
100.0.0.5 both have all the VPN rules in place on all the 3 nodes.

It could be possible that the rule came a bit late but the test attempted the ssh'ing earlier.

Patch-set: 25