Update patch set 1
Patch Set 1:

(2 comments)

First of all, my comments here are not specific to this change.

There is an abuse of classes with static methods throughout the whole repository and I think that we should tend to reduce them instead of adding more. 

There is an ongoing effort for Dependency Injection with the blueprint stuff and all, and, in my opinion, having static methods implementing business logic is a bit like going on the opposite direction. 

I'm not saying that static methods should be banned forever, it's just that my feeling is that we are misusing them.

Patch-set: 1
Reviewer: Gerrit User 5350 <5350@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0