Update patch set 1
Patch Set 1: Code-Review-1

Vishal,

Yeah we want OFTunnels in Carbon, but let's hold off on this a bit.
I think a configurable flag will be good to have - with the default being false until we have actually tested OF based tunnels.

What are your thoughts on the lack of BFD monitoring when we enable this? It seems it will get rid of all the TZ related issues, but are there plans to complement this with an alternative monitoring solution?

Patch-set: 1
Label: Code-Review=-1