Update patch set 2
Patch Set 2: Code-Review-1

(6 comments)

Vivek this is a followup to your mail discussion with Koby, it should be easier to discuss the actual change using this.
I put in some comments but it's still marked WIP.. you had some concerns about behavior change of learning of the first packet, but it seems in both old and new case the first packet won't be learned (originally after the sleep).

Patch-set: 2
Label: Code-Review=-1