Update patch set 3
Patch Set 3:

> "I put in some comments but it's still marked WIP.. you had some
 > concerns about behavior change of learning of the first packet, but
 > it seems in both old and new case the first packet won't be learned
 > (originally after the sleep)."
 > 
 > Original code uses a single ARP packet to learn the MAC-To-IP
 > Mapping.
 > The sleep is synchronous to the packet handler.  The next packet
 > won't be handled until the first packet is used to learn the
 > MAC-TO-IP mapping.
 > 
 > However, the current review would require multiple ARP packets to
 > do the same learning.

Vivek I think we're misunderstanding something here.
In both functions when there is a sleep(2000), after the sleep the packet is not learned. The learning seems to be in the else.
What code makes the original ARP wait 2 seconds and then be used to learn?

Do you have a better suggestion on how to resolve this sleep?

Patch-set: 3
Label: Verified=0