Update patch set 3
Patch Set 3:

"Vivek I think we're misunderstanding something here.
In both functions when there is a sleep(2000), after the sleep the packet is not learned. The learning seems to be in the else.
What makes the original ARP wait 2 seconds and then be used to learn?
Do you have a better suggestion on how to resolve this sleep?"

OOPS I overlooked that add is not right below the Thread Sleep.
I had some mail exchanges on this with Koby which I forwarded.

The sleep is required because of movement of adjacency for this IP Address from one vpn interface to another... This movement would result in churn of the vrfentry for this prefix thereby resulting in a situation where we will hold a released label for this prefix.

A better solution (definitely not the best) to this problem is to have a listener on vpn-interface Op adjacency removals and have that listener trigger clearance of the dirtyIPMap that is being constructed..
Basically the removeMIPAdjacency will trigger adjacency removal from VPN Interface Op Data Store.. After that Adjancency removal event comes back, we process the equivalent ip address from the dirty ip map.

Until then the dirty ip map can continue to hold the latest MAC Address resolved for that ip..

Patch-set: 3