Update patch set 4
Patch Set 4:

Alon had a question on why we learn with with ARP requests and not just GARP and ARP responses only..  

For this better to go to RFC here for ARP: https://tools.ietf.org/html/rfc826

We are supposed to learn only from GARP requests and ARP responses.

It was coded to learn from ARP requests initially because in environments where a GARP is not generated for migrated IPs, but a traffic is already initiated from such migrated IP, it will enable the controller to learn the IP to MAC mapping at the first possible opportunity.

That said, the other communicating end in order to populate its ARP table will anyway do an ARP request again towards the migrated IP and that will entail an ARP response from the migrated IP which is again learned..

So to conclude, we can remove the logic of learning from ARP requests and retain learning only via GARP and ARP responses.

Vivek

Patch-set: 4
Label: Verified=0