Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

Very sorry for the delay in getting to review this one. I love the clean-up work done in this Gerrit, and want to +1 it to get this in.  Giving -1 just because of a something I spotted in AbstractDataChangeListener that's worth double checking IMHO.  Also, as you know, personally overall I'd like us to push even further than what you're doing here, and in https://git.opendaylight.org/gerrit/#/c/47378/ am proposing to change a lot of @Nullable return types to Optional, to start with in aclservice (your changes here are for more projects); thus we have some conflicts.  But wanting to go further is not reason for this not to go in already first (or after mine; whoever comes 2nd adjusts the conflicts to the work of the other; doesn't matter), so after we clarify the doubt in AbstractDataChangeListener, I would +1 this if you would +1 c/47378, how about that? ;-)

Patch-set: 7
Label: Code-Review=-1