Update patch set 3
Patch Set 3: Code-Review+1

Hello.

We have found a bug in the cloudservicechain module. It should be merged in stable/boron SR2 also.

This fix has already tested so we can guarantee it corrects the issue. 

As the change is really light, we hope it can be into boron-SR2 release.

BR.

Patch-set: 3
Label: Code-Review=+1