Spec for NetVirt QoS
This spec file describes the changes to be made
in the Carbon release to complete the implementation
of Neutron QoS API support for NetVirt.

QoS Policies with Bandwdith Limiting Rule support was
added in Boron.  Carbon will add support for the DSCP
marking rule and for reporting supported QoS rules
back to Neutron.

Change-Id: I7e10b702fbbe9eb214336f500c00a579cb441e37
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>