Update patch set 4
Patch Set 4:

Thanks Peri, I will merge this spec as it seems there are no more comments for now.

A small note - we will be using hard_timeout and not idle_timeout at all, since we do want the packets to be re-punted to the ODL within a reasonable time if for some reason ODL did not apply a new flow for the MAC in table 50.

This hard_timeout will only be on the table 49 auto-learned flows.
The table 50 default flows will always be there and have no timeout.

Could you please take a look and review the actual implementation?
https://git.opendaylight.org/gerrit/#/c/48834/
https://git.opendaylight.org/gerrit/#/c/48835/

Patch-set: 4
Label: Verified=0