Update patch set 2
Patch Set 2:

> (1 comment)

This is exactly the same as what Venkat did for legacy (linked above).
The "0" represented a flags bitmask that was all off.
Now we want to turn on the second bit, to turn on "delete_learned" flag.

This means when the flow is deleted (i.e. when SG is deleted), all the flows automatically "learned" by this flow will also be deleted (resulting in no connectivity as expected). This is a good fix for boron as well.

We probably need a test-netvirt-learn-current by the way (or even move default to learn).

Triggering test-netvirt-current even though it doesn't really test this.

Patch-set: 2