BUG 7305: DHCP fails for Dual stack ports
For VM ports configured as dual-stack i.e. having both IPv6 and Ipv4
addresses, DHCP Service fails to resolve DHCP.

Issue is with current code that assumes IPv4 only and returns the first
fixed-ip. If IPv6 is the first IP, it doesn't work. Fix is to find and use
the first IPv4 address.

Change-Id: I3a9baa6f3e8fd675c07423050ff44630417cf414
Signed-off-by: Vishal Thapar <vishal.thapar@ericsson.com>