nat split-horizon changes.
Description : In one specific case, we have we have SNAT rules configured
on OVS1 and one floating-ip and one non-floating-ip vm on OVS2. Because of
split-horizon, the packet gets dropped on OVS1. Changes done to disable
split-horizon-protection on all TEPs(for type vxlan) on NAPT switch

Changes done in NAT to handle following use-cases.

Case 1 : Few of vxlan ITM Tunnels were already configured on the DPN which
gets elected as NAPT switch for an VRF
Solution : As part of SNAT flows installation, identify all the vxlan
interfaces configured on the switch and update the interface to set with
“override-split-horizon-protection”=true configuration. I believe
internally(ITM/interfaceManager) takes care of updating the already
existing table=0 flow for these interfaces with the addition of new
action(load:0->NXM_OF_IN_PORT[]).

Case 2: If an ITM tunnel configured post on elected NAPT switch
Solution : NAT needs to have an Listener which takes care of updating the
interface to set with “override-split-horizon-protection” configuration.

Case 3: Elected NAPT switch goes down and a new NAPT switch is re-elected.
Solution : This will be similar to Case1. Along with this, it is required
to remove back “override-split-horizon-protection” configuration on all
the vxlan interfaces of old NAPT switch.

Change-Id: I3b7d21b24ab3898a366a5121f9558f9ae7292e9e
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>
Depends-On: I2a93cdeef6f1dcd2b4d8b95315a7106970572067