BUG7539: duplicate local broadcast group backets
In a setup with more than two switches, when creating a network with one
port on each switch, sometimes the remote broadcast group has more than
one bucket directing to the local broadcast group.
because the list was initiated outside the for loop of the switches, the
local brodcast bucket was added to it more than once, hence the
duplication.

Change-Id: Ifa1decd5b238f45f97770e03eb10bd781c24cbfc
Signed-off-by: David Goldberg <gdavid@hpe.com>