Update patch set 2
Patch Set 2:

(1 comment)

Vivek, this code is for the following:

when the InterVpnLink is created, it must be checked if there is any router with an static route pointing to one of the InterVpnLink's endpoints. If so, we must ensure that the router belongs to the VPN *opposite* to the endpoint. 

When a route like this is found, we must retrieve the VPN to which the Router belongs and check that it is the opposite one (opposite to the endpoint/nexthop).

To speed things up I'm building a map (router->vpn) that is queried whenever one of this static routes is found. 

An alternative is to get rid of this map an consult VpnMaps for every route that is found. I'm now more inclined to this approach

Patch-set: 2
Reviewer: Gerrit User 5350 <5350@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0