Bug 7229: Allow certain ICMPv6 NDP packets by default
This patch programs OVS flows to allow ICMPv6 egress packets (from VM POV)
of type 133(RS), 135(NS) and 136(NA). Prior to this patch, these packets
were going through the conntrack module (in stateful implementation) and
were getting dropped as there was an inconsistency in the ct_state between
kernel and user-space conntrack implementations. After discussing this issue
with OVS team, the recommendation was to bypass conntrack for the ND packets.

Change-Id: I1a74046056bd683a71237c7d8ab8ab4606f29a96
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>