Update patch set 4
Patch Set 4: Code-Review+1

Peri, thanks for handling this! Even though this might look like just an error handling change - it might actually solve the failures we're seeing in CSIT runs.

Apparently those errors are happening on external network delete inside InterfaceRemoveWorkerOnElanInterface. Therefore, the job is retried several times, and blocks the DjC queue for several seconds. This delays the VPNInterface deletion, and it turn causes it to fail (because the interface state is already missing). When a new network is created, it fails because the last one was partially removed.

I think this should be cherry-picked into stable/boron, as it should improve the CSIT stability (at least the reports I was looking at), and it is not really risky.

Patch-set: 4
Label: Code-Review=+1