Bug 7436: Handle VpnInterfaces of VpnInstance
The problem:
VpnOpDataSyncer.waitForVpnDataReady is using java wait method
that does not return the thread to the pool while waiting.
If there are more compute nodes then CPU threads, all threads
in DjC will wait and there will be not availabe thread to
execute the code which waitForVpnDataReady waits for.

The solution:
VpnInterfaceManager holds all vpn interfaces that cannot be processed till the
relevant vpn instance in ready.
Updated the synchronized lock blocks to user vpnInstanceName instead of this.

Added cleanup of removed interface

Change-Id: I66da144356e17e778f1d04e5720befbece565fd5
Signed-off-by: Tomer <tomer.pearl@hpe.com>