Bug 7448 - External network recreation fails in newton nodl v2
Network delete in newton/networking_odl v2 resets some provider_attributes
prior to delete using a network-update.
This patch handles that update() by dealing with removing vpninterfaces on the external network.
For the case of network delete no new elements will be deleted as the physical network
is not configured

Change-Id: I4932610316c9445934334bb2211378b507b018f7
Signed-off-by: Tali <tali.ben-meir@hpe.com>