Use OVSDB external-id for Neutron port parent ref
This patch adds logic when neutron port is added, to attempt to find
an OVSDB termination point, containing an external-id that points to
this port, and set it as that interface's parent ref.

If no such termination point is found, we no longer do the previous
behavior of trying to guess the parentRef by the VIF type - the parentRef
will be left empty until such a termination point is added.

Depends-On: https://git.opendaylight.org/gerrit/49906
Depends-On: https://git.opendaylight.org/gerrit/49907
Change-Id: I86f1f964c363af982c2460d40ef6b7afb7f56e59
Signed-off-by: Alon Kochba <alonko@hpe.com>