Bug 7476: Configure Reachable Time in IPv6 Router Advt
Currently ODL IPv6 Router does not specify the "Reachable Time" in the
Router Advts. So, Neighbor unreachability detection algorithm in the
hosts use the default value of REACHABLE_TIME (i.e., normally 30 seconds)
which is distributed with a random factor between MIN_RANDOM_FACTOR (.5)
and MAX_RANDOM_FACTOR (1.5).

In this patch, we configure "Reachable Time" to two minutes to reduce the
number of Neighbor Solicitation messages in the tenant network. In a future
patch, we can make this a configurable value.

Change-Id: Icd40d51d9ec9e34306c1d05fdaa7348fb25ad302
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>