Update patch set 1
Patch Set 1:

Vishal, Alon, the code should aready create the ports for provider nets even if the bridge already exists. My point is that the procedure needs to be re-entrant and state/config agnostic. If there is something we're worried will get screwed up if the bridge already exists, it needs to be handled the same was as provider networks.

Patch-set: 1