Update patch set 1
Patch Set 1:

Hi Sridhar,

Thanks for the update. I do see handling in the code not to create tunnels for elans which are not vxlan. However, I couldn't find anything for routers connecting non-vxlan networks.

In any case, could you please write a review comment there with the exact configuration where it failed, so it'll be handled (even if we merge this current patch, it'd be overridden by the other one once merged - so I want the other patch to be fixed as well).

Patch-set: 1