Bug 7530 : ElanPacketInHandler mutexes are too coarse
Mutex for manipulating flows related to MAC was locking on
ELAN + MAC, when could lock ELAN + MAC + DPID. By locking
on the finer-grained mutex, we can support much bigger scale.
Problem occured in scale testing when 180 computes were
getting packet ins with the same MAC (External Network), and
every OVS was competing with all other OVSs for the lock of ELAN+MAC.

ElanPacketInHandler also updates a global data structure
of MACs that must be locked with the coarser mutex of
ELAN+MAC. This is why this patch splits the ElanPacketInHandler
into 2 jobs instead of one.

The transactions there also caused OptimisticLockException because
delete was called and submitted and then put was called and
submitted. This can't work due to the asynchronoush nature of
MD-SAL. There is no need in delete anyway if you call put, so
the delete was removed.

The 10 seconds protection was removed because it is unnecessary
as it already exists in the OVS with a learn rule.

The MAC migration code never worked because the condition could
ever be true there (it's a bug). I preserved the same behavior.

Change-Id: I1aa038f9e0d271b94ffee6076654571b2e811d37
Signed-off-by: Guy Sela <guy.sela@hpe.com>