Update patch set 8
Patch Set 8:

(1 comment)

Hi Tomer,
I found this commit by chance.
Mostly agree with the change but, why dont we take a subscriber/publisher approach?

VpnInstanceListener does not need to know which classes are involved. By specifically telling VpnInstanceListener that VpnInterfaceManager must be notified, we are introducing a coupling that is not really needed. 

But mainly because with the subscriber-publisher approach, InterVpnLink can benefit too.

Patch-set: 8
Label: Verified=0