Update patch set 1
Patch Set 1:

I think there is a bigger issue here. Is this thrown only from newton?
If it does, it sounds similar to https://bugs.opendaylight.org/show_bug.cgi?id=7448 - network delete starts with an update that reset most fields including network-type followed by the actual network delete
If we want to support this as-is and this is not a neutron bug than the actual code called from NeutronNetworkChangeListener.remove() needs to be called from update() when network-type is null
Does this make sense to support this or this is neutron bug?

Patch-set: 1