Update patch set 1
Patch Set 1:

> (3 comments)
 > 
 > can you elaborate on why specific lldp packets punting necessary ?

Hi Peri, please take a look at [1] as well. Currently L2 service is forwarding packets that must not be forwarded - they should be either be handled or dropped.

Since there is LLDP handling in ODL (in openflowplugin) we added the ability to punt those. We didn't want to enable this by default, as we're not sure about performance implications yet.

[1] https://git.opendaylight.org/gerrit/#/c/50481

Patch-set: 1
Label: Code-Review=0