Update patch set 3
Patch Set 3:

> Okay Koby, I understand. But should we use ELAN known DMAC table to
 > punt LLDP packets ? Why can't we use table 0 itself to punt lldp
 > packets with flow entry matching on eth_type (0X88CC) ?

We could, but I think the proper place to do that is inside the L2 service?
As I see it, the L2 service must make sure it does not forward packets that must not be forwarded by L2 switches. Also if it implements specific L2 protocols (like LLDP), it should be there as well?

Patch-set: 3
Label: Verified=0