Bug 7567: External subnet group is not updated with external gwmac
The external subnet group is not getting updated even though
the resolution for external gateway macaddress is successful.

The fix here is to make the the external subnet group programmer
to listen for Learned IP-MAC datastore which would really hold
the external-subnet-gateway-ip TO external-subnet-gateway-mac
mapping.

Also multiple external subnets can share the same subnet ip addresses
and so this patch fixes that too..

Change-Id: Id8c2ba8ace8dbeaf20ceb35d81f43a07e50c3a9f
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>