Update patch set 3
Patch Set 3:

> Yes Koby, I get your point. But with current netvirt pipeline, to
 > reach ELAN pipeline the packet has to go all over through ACL, DHCP
 > and L3VPN. So I felt it is unnecessary. Also I think LLDP handling
 > code in ODL doesn't leverage anything from ELAN (i.e. from elan tag
 > and lport tag).

Yakir, please see comments on https://git.opendaylight.org/gerrit/#/c/50481 to combine logic (you can depend this patch on the other)

Peri, looking at the two patches, it seems like there will be a list of multicast addresses with each having a drop or punt logic.
LLDP is the first example of one with punt handling, and the rest are dropped in the other patch.

Genius could handle this if we were to decide all projects need to have this specific L2 behavior, however I'm not sure that's right, as other applications might want different behavior.
If we do it netvirt it seems like the ELAN service is the natural place for it..

Patch-set: 3
Label: Verified=0