Bug 7601: Cleanup Elan instances when a network is deleted
Currently, prior to a network delete event, an update event would be
received which removes the augmentation for networkType. So, in delete
event, when we try to check for the networkType, it would be empty and
there is also an exception, because of which elanInstances are not
getting deleted. In this patch, we avoid the networkType check itself,
since the network itself would not be created in the first place if its
not a supported networkType.

Change-Id: I349a51f4df666ef2c2912ee552f159b6af16de76
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>