Update patch set 1
Patch Set 1:

> there seem to be blueprint issues.

ah yes, as-is, this leads to a cyclic dependency between ElanUtils and HwvtepTerminationPointListener ... :-( The BP error handling in such cases is REALLY BAD - e.g. Spring or Guice DI have clear message for such cases; BP seems to just "fail" - hopefully later versions of Apache Aries' Blueprint in Karaf 4 are better...

This would need more work.  Not a top prio. for me right now; I'll may be get back to it some time later.  If anyone else wants to have a go at how to refactor the code to avoid the static without a cyclic dependency, please do feel free to jump on this!

Patch-set: 1