Update patch set 1
Patch Set 1:

Hi Vivek.
Removing the ARP responder rule done  in processVpnInterfaceDown(). If you take a look at this function, you'll see that VRF entry is removed only if next hop list of this VRF is empty. 
So the rule is deleted while vm reboot, but VRF entry remains because of null string in the next hop address list. The ARP responder rule is installing while vpn interface adjacencies treatment inside of the method processVpnInterfaceAdjacencies(), but it checks if related VfrEntry already exists in FibTable. If it does then nothing’s done...

Patch-set: 1
Label: Verified=0